Use explicit configuration in /etc/default/beast-splitter, don't hotplug.
The autostart-on-device-hotplug was interesting but tricky to get the right
config and still work with network input. So move to an explicit
configuration model where you need to set up /etc/default/beast-splitter.

This also adds a wrapper script that handles the case where the defaults
file says not to run beast-splitter at all.